demo/requirements.txt
ingredient_parser==1.0.2
ingredient_parser_nlp==0.1.0a4
numpy==1.21.5
pandas==1.4.4
Pillow==9.4.0
python-dotenv==0.21.1
requests==2.28.1
streamlit==1.17.0
